ref: create shared lib for sample apps
