+++ Samples/SentrySampleShared-Swift/SentrySampleShared-Swift/DSNStorage.swift
@@ -4,10 +4,9 @@ import Sentry
 /**
  * Stores the DSN to a file in the cache directory.
  */
-class DSNStorage {
-    
-    static let shared = DSNStorage()
-    
+public class DSNStorage {
+    public static let shared = DSNStorage()
+
     private let dsnFile: URL
     
     private init() {
@@ -17,12 +16,12 @@ class DSNStorage {
         dsnFile = cachesDirectory.appendingPathComponent("dsn")
     }
     
-    func saveDSN(dsn: String) throws {
+    public func saveDSN(dsn: String) throws {
         try deleteDSN()
         try dsn.write(to: dsnFile, atomically: true, encoding: .utf8)
     }
     
-    func getDSN() throws -> String? {
+    public func getDSN() throws -> String? {
         let fileManager = FileManager.default
         
         guard fileManager.fileExists(atPath: dsnFile.path) else {
@@ -32,7 +31,7 @@ class DSNStorage {
         return try String(contentsOfFile: dsnFile.path)
     }
     
-    func deleteDSN() throws {
+    public func deleteDSN() throws {
         let fileManager = FileManager.default
         if fileManager.fileExists(atPath: dsnFile.path) {
             try fileManager.removeItem(at: dsnFile)
