+++ Samples/iOS-Swift/iOS-Swift/AppDelegate.swift
@@ -1,3 +1,4 @@
+import SentrySampleShared_Swift
 import UIKit
 
 @UIApplicationMain
