+++ Samples/iOS-Swift/iOS-Swift/Profiling/ProfilingViewController.swift
@@ -1,4 +1,5 @@
 import Sentry
+import SentrySampleShared_Swift
 import UIKit
 
 class ProfilingViewController: UIViewController, UITextFieldDelegate {
