+++ Samples/iOS-Swift/iOS-SwiftClip/AppDelegate.swift
@@ -1,4 +1,5 @@
 import Sentry
+import SentrySampleShared_Swift
 import UIKit
 
 @main
@@ -16,7 +17,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
             options.tracesSampleRate = 1.0
             options.sessionTrackingIntervalMillis = 5_000
             options.initialScope = { scope in
-                scope.injectGitInformation()
+                injectGitInformation(scope: scope)
                 return scope
             }
